TCCP: Feature flag logic to support beta testing
Currently the TCCP feature flag logic prohibits it from running in
either beta or production. Soon we want to be able to test on beta.

This change modifies the feature flag logic such that TCCP is always
available locally or on our internal DEV4 testing server. Deploying it
in other environments will require manually enabling it via the
Wagtail UI.