TCCP: Add size filtering
This commit adds a new `top_25_institution` field to the
CardSurveyData model, indicating whether the card's institution is one
of the 25 largest issuers.

This new field is then used to power a new "Small institution" checkbox
on the list view filter.

This field does not exist in the most recent TCCP dataset (1/1-6/1/23),
so it needs to be added after the dataset Excel has been loaded. To do
so, a new `patch_tccp` command has been added that takes as input a
text file containing a plain list of the top institutions, which must
match exactly the institution names in the dataset.