Remove unnecessary redirect in YFPTG
The "Your financial path to graduation tool" tries to load school data
from the API using a URL like this:

/paying-for-college2/understanding-your-financial-aid-offer/api/school/1234

Note the lack of trailing URL. These API URLs are always served with a
trailing URL, so the URL should be

/paying-for-college2/understanding-your-financial-aid-offer/api/school/1234/

As it happens, Django will currently respond with a 301 redirect to
the right URL, but this is an unnecessary delay and page load. Removing
this unnecessary redirect will speed up our functional tests for this
page, not to mention the end user experience.