TCCP: Add 3-tier card rating based on purchase APR
With this change, cards in the TCCP dataset are given a rating
indicating how their purchase APR compares against all other card
purchase APRs within the same tier. On the frontend this is rendered
as a 3-tier system: pay less interest, pay average intereset, pay more
interest.

The actual calculation used to place cards into tiers is based on the
percentile of the tier-specific purchase APR:

- if the card APR < 25th percentile of all tier-specific purchase APRs,
we label it as "pay less interest"
- if the card APR < 50th percentile of all tier-specific purchase APRs,
we label it as "pay average interest"
- otherwise, we label it as "pay more interest"

This change adds a key to the card listing page to help users understand
the three tiers.

On the backend this works by first computing summary statistics across
all cards (for each tier), including the 25th and 50th percentiles. It
leverages the tailslide [0] Python package for this, which under the
hood uses the PostgreSQL PERCENTILE_CONT function [1].

To test, run a local server and visit:

http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/

Alternatively, visit:

http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/?format=json

to see the same data as JSON; note the "stats_all" key which contains
the summary stats across the whole dataset, including the percentile
cutoffs.

[0] https://github.com/ankane/tailslide
[1] https://www.postgresql.org/docs/14/functions-aggregate.html