TCCP: Multiselect situations + situation content
This comment modifies the TCCP landing page so that users can select
more than one situation. It also creates the framework for each
situation to have two templates: one for the landing page and one for
the list view. Situation list view content will be shown above the list
view results, depending on which situations the user selected.