TCCP: Exclude non-targeted cards
Some credit cards offer a purchase APR for a specific credit tier but
are not marked as being "specifically targeted" to that credit tier,
recorded in the TCCP survey under the "Targeted credit tiers" column.

Previously TCCP filter results would include these cards, so if a user
marked their credit tier as 620-719, they might see cards that, while
they offer a purchase APR for this range, are not marked as being
"specifically targeted" to this range.

This change removes those cards from the results.