TCCP: JSON detail view + nicer enum display
This commit refactors the TCCP detail view so that it is accessible via
JSON, at, for example with the current dataset at
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/baxter-credit-union-travel-rewards-visa-signature/?format=json

This work was done in service of making it easier to debug this view and
to provide a more consistent framework for converting TCCP enum values
to user-friendly strings on the frontend.

The list view still also works as JSON, at
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/?format=json

This change includes two places where we now use more user-friendly
strings: card rewards, and card state names.

To see the more concise rewards labels on the list view, visit
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/
and notice the shorter labels under "Rewards".

To see the nicer labels for both rewards and states, visit a card detail
page with both of these, for example
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/banner-bank-banner-bank-trurewards-mastercard/