TCCP: Modify card purchase APR logic
This change modifies the logic for which cards are included in TCCP
filtering results by expanding the purchase APR logic. Previously, we
would only include cards that had a tier-specific purchase APR, for
example the `purchase_apr_good` column in the TCCP dataset.

With this change, we may also select a purchase APR from other columns
in the dataset, for example the `purchase_apr_median` column. We may
also use the `purchase_apr_min` and `purchase_apr_max` columns to
define a range of APRs, if necessary.

See the changes to cfgov/tccp/models.py for extensive documentation on
the new logic.

To test, run a local server and explore card results under
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/.
Results now contain (occasional) examples of cards with a range of
credit scores. You may also want to visit the JSON version at
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/?format=json
to see the summary statistics being used.

If you click through to individual card detail pages, you'll notice that
we are now using non-tier-specific purchase APRs to populate the list
view.

Additionally, the detail view now includes summary statistic logic to
populate future display of card ratings there. Visit a detail page in
JSON view to see the new columns and summary statistics, for example
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/first-community-credit-union-visa-platinum/?format=json.

See internal
https://github.local/Design-Development/Design-Thinking-and-User-Research/issues/270
for additional context and discussion.