TCCP: Display account fee as a number
With this change, the TCCP list view now displays a dollar amount for
each card's "account fee". This value is based off the values of several
different columns in the TCCP dataset:

- periodic_fee_type: what type of periodic fees a card has, which can
have zero or more values of "Annual", "Monthly", "Weekly", and "Other"
- annual_fee: amount in dollars of annual fee
- monthly_fee: amount in dollars of monthly fee
- weekly_fee: amount in dollars of weekly fee

If a card has any fee type besides "Other", we can estimate the annual
fee by multiplying the monthly fee by 12 and the weekly fee by 52. If
a card has an "Other" fee, we can't do that, because it isn't
represented in a predictable way in the various related "Other fee"
columns, for example "Other Periodic Fee Name".

For this reason, if a card has an "Other" fee, it will show up as
"See details" (we will iterate on this). If a card has no fee, it will
show up as "None". If a card has a value that can be computed, we show
that dollar amount.