TCCP: Card details page template consolidation
This commit consolidates some of the Jinja logic in the TCCP card
details page. It creates a new `apr_table` macro to remove some of the
duplication in the current details page. It also avoids the need to
hardcode the various credit tiers in the template, instead reusing the
values from the backend.

There is more cleanup and consolidation to be done on this page, and
this is just a first iteration.

To test, run a local server and visit cards under
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/.

Past PRs iterating on this page are a good place to see example URLs
to test:

https://github.com/cfpb/consumerfinance.gov/pulls?q=is%3Apr+author%3Aniqjohnson+is%3Aclosed