Convert DB unique index to unique constraint
This is a one-off fix to convert an existing unique index on a Wagtail table into a unique constraint. See internal Design-and-Content-Team#389 for a description of the problem with a different constraint, https://github.com/cfpb/consumerfinance.gov/pull/8066 for the fix this is
based on.

The mode QueryDailyHits is and its `UNIQUE` constraints are removed in [`wagtail.search.migrations.0008_remove_query_and_querydailyhits_models.py`](https://github.com/wagtail/wagtail/blob/stable/6.0.x/wagtail/search/migrations/0008_remove_query_and_querydailyhits_models.py) in Wagtail 6.0. This migration will prepare our database for the constraint's constraint's removal.