Make Contact phone numbers clickable links
This commit alters our template for the ContactPhone molecule such that
phone numbers are rendered as clickable links.

The regular phone number and TTY phone numbers are both already
required to be a string containing only numbers, which complies with
the RFC for tel: links [0]. The vanity number may contain letters,
which are not supported by the RFC, and so is not made into a
clickable link.

As part of this change I removed the bolding we currently have for
phone numbers to make the weight more consistent with the rest of the
text.

[0] https://datatracker.ietf.org/doc/html/rfc3966