TCCP: Cleanup detail page fee logic
This commit simplifies some of the logic on the TCCP card detail page,
removing code that tried to determine if a card has a varying periodic
fee. The dataset schema doesn't provide enough information for us to
reliably report this to users.

This change simplifies the fee rendering so that we only show annual,
monthly, and weekly fees, without information on periodic changes.

See internal
https://github.local/Design-Development/Design-Thinking-and-User-Research/issues/306#issuecomment-354469
for additional context behind this change.

To test, run a local server with the latest data and visit a card page
with fees, for example
http://localhost:8000/consumer-tools/credit-cards/explore-cards/cards/credit-one-bank-national-association-platinum-visa-for-rebuilding-credit/