Additional flake8-simplify fixes with ruff
This is the result of running `ruff check --select SIM --fix --unsafe-fixes [specific file]` on each of the files ruff flags with flake8-simplify warnings, and inspecting the results, and manually editing where necessary.

There are still some flagged issues after this commit, but for those the `--unsafe-fixes` are... not ideal.