Fix extra whitespace in some link markup cases
Our link icon markup incorrectly leaves behind whitespace in certain
cases. For example, given a link like this:

<a href="https://example.com"> text </a>

the markup will generate:

<a href="https://example.com" class="a-link a-link--icon">
  <span class="a-link__text"> text </span>
  <svg class="cf-icon-svg cf-icon-svg--external-link">...</svg>
</a>

The extra whitespace in the <span> gets underlined incorrectly.

See "View a webinar on the rule" on
https://www.consumerfinance.gov/know-before-you-owe/ for an example
where this happens on the live site.

This fix attempts to strip surrounding whitespace from the link text
to prevent this happening.