Remove broken and deprecated CFGOVPage unit test
This unit test was failing for me with certain invocations of Python
unit tests, for example:

tox -e unittest -- v1.tests.models.test_base

This fails due to a RecursionError in Django's settings code, whereas
running those tests in reverse passed:

tox -e unittest -- --reverse v1.tests.models.test_base

Digging into this, the problem is that this test patches the Python
super() call which has all sorts of potential side effects. In this
case, it was breaking something about how Django settings are cached
the first time they are accessed.

As it turns out, this test can be removed - it was created at a time
when our CFGOVPage.serve() method was sometimes invoked via AJAX, which
is no longer the case.