Cache SVG icons in memory
Currently any use of the {% svg_icon %} template tag loads the SVG
contents from disk. This commit modifies that behavior so that SVG icons
are cached in memory, to avoid the need to load from file each time.

In our icon library [0] icon files range from 220 bytes to 2.6 KB in
size. The total size of all icons is 2.8 MB.

[0] https://github.com/cfpb/design-system/tree/6b566bdb7ae2ce4c80a1cd05404bf3ea2e8fe6fa/packages/cfpb-icons/src/icons