Optimize JobListingTable DB queries
This commit optimizes rendering of the JobListingTable organism found on
https://www.consumerfinance.gov/about-us/careers/current-openings/ so
that it takes 5 database queries instead of 13.

When this code was last touched in April 2020, django-modelcluster didn't
properly support prefetching of ParentalManyToManyFields (see
django-modelcluster#101). In May 2020 I opened a PR to remedy this (see
django-modelcluster#122), and this was released in September 2020, but
I neglected to revisit this code to add the optimization. This commit
does so.

All existing tests continue to pass properly.