Add an "Icon" button to Draftail for SVG icons
This change adds a Draftail plugin for inline SVG icons. This is based
on [Wagtail's documentation for extending Draftail](https://docs.wagtail.org/en/stable/extending/extending_draftail.html) and on other Draftail plugins that we've explored (wagtail-draftail-anchors and wagtail-footnotes).

Draftail plugins manage content in two ways, generally:

1. DraftJS contentstate
2. Database state

For SVG icons, when a user adds the icon to the rich text editor, a
DraftJS Entity is created that wraps a single space (because Entities
must have >0 ranges), and that Entity gets rendered in the rich text
editor by our Draftail plugin decorator as an `img` tag that loads our
SVG icon from its static path. If the icon doesn't exist, the image will
appear broken. The space that the Entity wraps is discarded for rendering.

When the user saves the page, the DraftJS contentstate is converted to
HTML(-ish) and stored in the database. At conversion time, the Entity is
converted into a `<span data-icon-name="icon-name">` and the contents of
the SVG is loaded via our `svg_icon()` function and stored inside that
`span` tag The space that the Entity wraps is discarded for rendering..

When the page is loaded back into the rich text editor, any
`span[data-icon-name]`-matching tags are converted back to DraftJS
Entities that wrap a single space, which the Draftail plugin decorator
will convert to an `img` tag, etc.