TCCP: Fix phone number parsing
This commit modifies how phone numbers are rendered on the frontend of
the "Explore credit cards" tool. It fixes an issue with parsing phone
numbers that have spaces in them in the TCCP dataset. The TCCP dataset
contains cards that only provide a single phone number, but sometimes
this can include spaces, for example:

1 (800) 555-1234

Instead of parsing this like multiple phone numbers separated by spaces,
we need to parse this as a single phone number.