Replace deprecated user form config
`WAGTAIL_USER_CREATION_FORM` and `WAGTAIL_USER_EDIT_FORM` are deprecated in favor of subclassing Wagtail's `UserViewSet` class and providing a custom `WagtailUserAppConfig` class that overrides that of `wagtail.users`. See https://docs.wagtail.org/en/latest/releases/6.2.html#deprecation-of-wagtail-user-edit-form-wagtail-user-creation-form-and-wagtail-user-custom-fields-settings

This is... not my favorite bit of Django code, and feels a little anti-idiomatic. Effectively what happens here is that the `LoginUsersAppConfig` that we define (alongside `LoginConfig` — it doesn't work when combined) both allows us to override `user_viewset` and provide our own ***AND*** replaces that of `wagtail.users`, so that `login.apps.LoginUsersAppConfig` masquerades as `wagtail.users` for the sake of Django's app registration. But this is the direction Wagtail has gone for user and group viewset overrides 🤷🏻‍♂️.