Remove local footnote custom rendering
This updates wagtail-footnotes to 0.12 and removes our custom `RichTextBlockWithFootnotes` subclass of wagtail-footnotes own `RichTextBlockWithFootnotes`, tests, and switches to using the upstream `RichTextBlockWithFootnotes`. 0.12 includes our changes in https://github.com/torchbox/wagtail-footnotes/pull/70, which were based on this code being removed.

This change updates `RichTextBlockWithFootnotes` in-place in old migrations. Removing `v1.blocks.RichTextBlockWithFootnotes` requires these migrations to be updated, and because there are no schema changes, new migrations are not required.