Freeze before-you-claim tests in 2024
We mock the responses from the retirement API, however the starting year was relative t the current year. This change freezes the starting year to 2024 and adjusts the month to ensure the tests pass against the fixture data.

Hopefully this avoids any future failures due to the continued march of time.