Fix (for real) the before-you-claim tests in time
#8552 froze the before-you-claim functional tests in time, but I missed a few things.

We're still a couple years out from when our API intercept date URLs were set to. This means we were testing directly against the API endpoint instead of getting a mock response. This was causing intermittent  failures.

I also think the "under 50" tests have probably been hitting the actual API endpoint as well for a while — there seems to be a mistake here in the intercept target date, 1969 instead of (now) 1959.