Fix Docker build failure due to missing sass
Trying to build the Docker image for this project following the
instructions [0] using this command:

```
docker build . -t your-desired-image-name
```

results in an error due to "Cannot find package 'sass'". This is because
the use of "production" for FRONTEND_TARGET in frontend.sh [1] does a
`yarn install --production` which excludes devDependencies, which
is what pulls in sass [2].

To fix this, I've modified the Dockerfile to remove use of
FRONTEND_TARGET and frontend.sh, and just call `yarn` and
`yarn run build` directly instead.

[0] https://cfpb.github.io/consumerfinance.gov/running-docker/#just-docker
[1] https://github.com/cfpb/consumerfinance.gov/blob/b63f58fd824eca2a17c099f9fd2606f1b38670b7/frontend.sh#L50-L53
[2] https://github.com/cfpb/consumerfinance.gov/blob/b63f58fd824eca2a17c099f9fd2606f1b38670b7/package.json#L39-L40