Fix link markup for certain non-cf.gov URLs
Currently certain archive-it.org URLs are missing an external link icon
due to an issue with the regex we use to detect non-cf.gov URLs.

For example, a URL like this:

https://archiving-service.com/1234/https://www.consumerfinance.gov/about-us/blog/fall-2015-rulemaking-agenda/

is missing its external link icon because the current regex matches the
"www.consumerfinance.gov" part of the URL, even though it's in the path
part, not the host part.

This commit fixes that and adds a regression test.

To test, run a local server and visit:

http://localhost:8000/about-us/blog/fall-2015-rulemaking-agenda/

and compare it against production:

https://www.consumerfinance.gov/about-us/blog/fall-2015-rulemaking-agenda/