Restore use of DATABASE_URL for DB configuration
Previously cf.gov supported use of the DATABASE_URL environment variable
to configure the Django database, as implemented by the dj-database-url
package [0]. This functionality was removed in 2022 to support a
since-abandoned migration to Helm [1]. Newer EKS/Helm workflows
allow us to use a single environment variable for DB configuration,
motivating restoration of this capability.

With this change, we no longer rely on environment variables like
PGHOST, PGUSER, etc. but instead rely on the single DATABASE_URL
variable.

See internal WAM#8 for additional context.

[0] https://github.com/jazzband/dj-database-url
[1] 4d0b24b1d326a884545a3a066ad765958e379541