Database script updates to support containerization
This commit modifies several scripts to make it easier to setup the database as part of a containerized cf.gov deploy.

It removes a portion of the refresh-data.sh script that attempts to manually rename the existing 'cfpb' schema in our database dumps, to a schema named for a different Postgres user. We no longer use the PGUSER environment variable as of 59b109b2a3d85fa551a8cb37b82cdbc9ad9058e8, so this code doesn't work right now.

In the future, if we need to import a 'cfpb' schema into a database where the username is something different (for example in an RDS instance where we might have less control over the username) we might have to revisit this code.

This change also consolidates the various 'DB setup' steps we do after loading a dump -- running migrations, creating the cache table, and running our 'initial' data script, reducing some duplication.