Fix incorrect Python tests using mock.called_with
This followup to PR 8711 fixes various Python tests that are incorrectly
trying to call `called_with` on a mock object, when the correct call is
`assert_called_with` [0].

Most of these tests are in the Paying for College disclosures code, and
one is in iRegulations.

Currently these tests don't actually test anything, as
`assertTrue(something that is truthy)` does nothing.

[0] https://docs.python.org/3.8/library/unittest.mock.html#unittest.mock.Mock.assert_called_with