Fix: Properly handle posting of long GitHub alerts
When cf.gov requests generate errors with excessively long tracebacks,
we're not able to post them to GitHub issues.

GH doesn't seem to document an official limit but according to
https://github.com/dead-claudia/github-limits?tab=readme-ov-file#issue-description
the limit is 65536 characters. When we post excessively long issues to
GH, we get a 422 Validation Failed error, which prevents us from
recording them.

This commit truncates the stack trace so that we can at least record
the error message.

See internal DEVPLAT-1465 for additional context.