Fix: Better handling of errors on 500 page
Currently if our 500 error page itself triggers another error (say, if
the connection to the database is down, and the error page tries to
access the DB to query flag state), we fall into an infinite recursion
of error handling.

Ideally our 500 error page should be static, and not access the
database or other resources that might trigger another error.

As a more immediate fix, we can change 500 errors to a simpler text
response if we're not able to render the proper 500 page. Other errors
(like 404s) should still be able to trigger the 500 page if their
own error page (like the 404 page) fail to render for whatever reason.