Fixes and tests for the cfgov-apache image
This commit adds tests for the cfgov-apache image that will be part of
the cf.gov EKS deployment. This image is not used in production yet.

The CFGOV_APPLICATION_HOST variable is changed to
CFGOV_APPLICATION_PROXY which allows for specifying the cf.gov Django
container more flexibly (e.g. http://app) instead of forcing that
container to always run on port 8000.

The cfgov-apache Dockerfile has also been modified to give the Apache
user (www-data) permission for the /usr/local/apache2/logs directory,
as this is where Apache writes its pidfile. Although in EKS in practice
we will use a readonly filesystem and mount a volume for this purpose,
it should still ideally be possible to run the container without
always having to do this. This www-data permission change makes this
possible.

Finally, this commit adds a new GitHub Action that builds and tests the
cfgov-apache image when there are changes to the cfgov/apache
directory. This will help prevent regressions in future.