Simplify print style for links with URL text
Currently the site print style marks up links with their URL, for
example, given a link with text "Example" that points to example.com,
the link will render like

Example (example.com)

This is overly verbose in cases where the link text is actually the
same as the URL, in which case you get duplication, like

example.com/foo (https://example.com/foo)

This commit attempts to modify this markup logic to avoid this
duplication, while keeping existing behavior for all other links.

This is implemented by adding a new class on the backend to links
with URL link text, and having our frontend CSS hide the URL markup
if that class exists. It isn't possible to implement this logic in
CSS alone.

The backend logic attempts to be smart; for example, a link to
https://www.consumerfinance.gov/foo that has link text of
"consumerfinance.gov/foo" will still get the simplified styling.

See internal DeCo#641 for context.