Migrate wagtail-sharing config to settings
The 2.13 release of wagtail-sharing allows configuration via Django
setting, avoiding the need to run a management command for this purpose.
This simplifies deployments and configurations in multiple
environments.

This commit upgrades wagtail-sharing from 2.12.1 to 2.13, and migrates
the config to use a setting derived from the WAGTAILSHARING_HOST
environment variable.

This variable has already been populated on our internal deployment
repo. See internal DEVPLAT-1547 for additional context.