Deprecate SQS-based alerting
This commit deprecates the SQS-based alerting implementation that used
an SQS queue to log application errors and New Relic alerts to a
GitHub repository. This logic is being deprecated as part of the cf.gov
migration into a new AWS account.

All code under cfgov/alerts can be removed as part of this deprecation.