Optionally skip OpenSearch SSL cert validation
In 8794c48259309ff69286908655812b5bac67aec9 we set the OpenSearch DSL
coniguration setting "verify_certs" to always be true, requiring that
the OS SSL certificate is validated.

When running against a vanilla OpenSearch container (as we do in our
local Helm deployment configuration, maintained in our internal
cfgov-deploy repo), it's not possible to validate the default cert. This
setting causes OpenSearch connections to fail.

When running locally with a local OpenSearch container, it's okay to
skip cert validation.

This change makes it possible to disable OpenSearch SSL certificate
validation through use of the OPENSEARCH_DSL_SKIP_CERT_VERIFICATION
environment variable. If set, cert validation will be skipped. The
default behavior (for example in our deploys to EKS) is to continue
properly validating OpenSearch SSL certificates.