Push cfgov Docker images to GHCR
This commit updates the existing build-docker.yml GitHub Action so that
the cfgov image is pushed to GitHub Container Registry (GHCR) as a
public image for download.

The image gets pushed as ghcr.io/cfpb/cfgov/cfgov (using a "cfgov"
namespace to support other future related images).

This happens on both pushes to main and on pull requests.
Pushes to main are tagged with "latest" and "sha-<short-sha>".
Pull requests are tagged with "pr-<num>" and "pr-<num>-<short-sha>".

PRs against forks, merge group triggers, and manual triggers will
continue to build the image but will not try to push it.

Logic around building the cfgov-apache image remains unchanged.