Avoid DB query on search code import
Merely importing the search.elasticsearch_helpers module currently
invokes a database query to generate a list of search synonyms. These
synonyms are part of the Ask CFPB index. The DB query is run when the
code is imported, even though the synonyms themselves are only needed
when the index is generated in OpenSearch.

In Django 5.0+, this generates a warning, for example when running unit
tests:

> RuntimeWarning: Accessing the database during app initialization is discouraged. To fix this warning, avoid executing queries in AppConfig.ready() or when your app modules are imported.
>   warnings.warn(self.APPS_NOT_READY_WARNING_MSG, category=RuntimeWarning)

See release notes at [0], which say "Executing SQL queries before the
app registry has been fully populated now raises RuntimeWarning."

This commit modifies the way the synonym list is generated, making it
"lazy" such that the DB is not hit until the list of strings is
actually needed. This avoids hitting the DB on code import, and defers
it until the index is generated.

This behavior can be verified by populating the OS index and inspecting
it in Kibana with a query like "GET ask-cfpb/_settings" to see the full
list of synonyms.

[0] https://docs.djangoproject.com/en/5.2/releases/5.0/#miscellaneous