Stop checking out all branches in coverage action
In https://github.com/cfpb/consumerfinance.gov/pull/5922 back in 2020
we implemented a "temporary" fix in the backend coverage GitHub action
to checkout the full history of the repository.

This PR attempts to revert that to speed up the coverage check. There
shouldn't be a need to check out more than just the current branch.