Adapt mortgage-performance processing for Alto work flows
Alto poses some challenges to the mortgage-performance routine:
- Local operations can no longer post files to public s3 buckets
- Staging operations can't post directly to prod buckets
- Processing could be run in production, where exports would work. But
that exposes the mortgage pages to at least 15 minutes of interruptions
as tables are wiped and rebuilt and exports are generated.

To avoid that, we run processing in staging, where the results
can be verified by stakeholders before promoting the changes to prod.
The most efficient way to update the prod db is to load a table dump,
which takes a few seconds and saves having to re-process source data.

This patch offers options for test-processing new data locally and
dumping the public CSV files locally so they can be checked
or manually pushed to s3 if needed.

It also adds configuration so that the process will run in Alto staging,
where table dumps and public CSVs can be generated without affecting
production. Once a new data load is approved, we can push the table
dump and public CSVs to prod s3 for the final step.

Additions
Two env vars are added to facilitate downloading and the reading of the
source CSV, which lives in an internal GitHub repo. The new vars are:
- MORTGAGE_PERFORMANCE_SOURCE, which allows a management command to
fetch source data from the internal repo without exposing its URL.
This is now a required var for mortgage processing and has been added
to the common secret vars helm produces to deploy cfgov.
- LOCAL_MORTGAGE_FILEPATH, an optional var that will trigger a
local download of public CSVs to a path, instead of to s3.
If the var is missing, an s3 push will be attempted. If an s3 push fails,
it is skipped and noted in the logs, rather than disrupting the run.
- A function that derives the through-date from the source file's name.

These changes remove the need to figure out what the through_date should be,
and avoids the churn of moving the source file to s3 and then accessing
it again for processing.

Internal GHE documentation at mortgage-performance/wiki/Pipeline-steps
has been updated.